+++ src/drivers/device/i2c.h
@@ -33,7 +33,7 @@
 #pragma once
 
 #ifdef __PX4_NUTTX
-#include "i2c_nuttx.h"
+#include "nuttx/i2c_nuttx.h"
 #else
-#include "i2c_posix.h"
+#include "posix/i2c_posix.h"
 #endif
