+++ src/drivers/device/posix/vdev.h
